+++ packages/moonstone/Marquee/MarqueeDecorator.js
@@ -160,38 +160,28 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			};
 
 			this.invalidateMetrics();
-			this.checkMarqueeOnRender(props);
 		}
 
 		componentDidMount () {
-			this.initMarquee();
+			this.initMarquee(this.props.marqueeOnRenderDelay);
 		}
 
 		componentWillReceiveProps (next) {
 			const {marqueeOn, marqueeDisabled} = this.props;
-			this.checkMarqueeOnRender(next);
-			if (next.marqueeOn !== marqueeOn || next.marqueeDisabled !== marqueeDisabled) {
-				this.cancelAnimation();
-			} else if (!childrenEquals(this.props.children, next.children)) {
+			if (!childrenEquals(this.props.children, next.children)) {
 				this.invalidateMetrics();
+				this.cancelAnimation();
+			} else if (next.marqueeOn !== marqueeOn || next.marqueeDisabled !== marqueeDisabled) {
 				this.cancelAnimation();
 			}
 		}
 
 		componentDidUpdate () {
-			this.initMarquee();
+			this.initMarquee(this.props.delay);
 		}
 
 		componentWillUnmount () {
 			this.clearTimeout();
-		}
-
-		checkMarqueeOnRender (props) {
-			this.marqueeOnRender =	props.marqueeOn === 'render' &&
-									!props.disabled &&
-									!props.marqueeDisabled;
-
-			return this.marqueeOnRender;
 		}
 
 		/**
@@ -221,15 +211,33 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		/**
+		 * Checks to see if the children changed during a condition that should cause us to re-check
+		 * the animation state
+		 *
+		 * @returns {Boolean} - `true` if a possible marquee condition exists
+		 */
+		shouldStartMarquee () {
+			return	this.distance === null && !this.props.disabled && !this.props.marqueeDisabled &&
+					this.props.marqueeOn === 'render' ||
+					(this.isFocused && this.props.marqueeOn === 'focus') ||
+					(this.isHovered && this.props.marqueeOn === 'hover');
+		}
+
+		/**
 		 * Initializes the marquee by calculating the distance and conditionally starting the
 		 * animation
 		 *
+		 * @param {Number} delay Milliseconds until animation should start
 		 * @returns {undefined}
 		 */
-		initMarquee () {
-			this.calculateMetrics();
-			if (this.marqueeOnRender) {
-				this.startAnimation(this.props.marqueeOnRenderDelay);
+		initMarquee (delay) {
+			// shouldStartMarquee relies on a `null` distance to indicate the metrics have been
+			// invalidated so we have to calculate after this check.
+			if (this.shouldStartMarquee()) {
+				this.calculateMetrics();
+				this.startAnimation(delay);
+			} else {
+				this.calculateMetrics();
 			}
 		}
 
@@ -353,21 +361,25 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		handleFocus = (ev) => {
+			this.isFocused = true;
 			this.startAnimation();
 			forwardFocus(ev, this.props);
 		}
 
 		handleBlur = (ev) => {
+			this.isFocused = false;
 			this.cancelAnimation();
 			forwardBlur(ev, this.props);
 		}
 
 		handleEnter = (ev) => {
+			this.isHovered = true;
 			this.startAnimation();
 			forwardEnter(ev, this.props);
 		}
 
 		handleLeave = (ev) => {
+			this.isHovered = false;
 			this.cancelAnimation();
 			forwardLeave(ev, this.props);
 		}