+++ packages/spotlight/CHANGELOG.md
@@ -4,6 +4,15 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ## [unreleased]
 
+### Added
+
+- `spotlightDefaultClass` to `@enact/spotlight` export. Applying this class to an item in a
+	container will cause it to be the default spotted item in that container.
+
+### Changed
+
+- Spotlight containers now default to focus last selected item when gaining focus.
+
 ## [1.0.0-alpha.2] - 2016-10-21
 
 This version includes a lot of refactoring from the previous release. Developers need to switch to the new enact-dev command-line tool.